Split tests for target-lines into separate file.
